+++ Examples/gpu/UltralightNet.Vulkan.TestApp/Startup/Utils.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using Silk.NET.Core;
